+++ monkeytype/compat.py
@@ -7,7 +7,7 @@ from typing import Any
 
 try:
     # Python 3.7
-    from typing import Union, _GenericAlias  # type: ignore
+    from typing import Union, _GenericAlias, ForwardRef  # type: ignore
 
     def is_any(typ: Any) -> bool:
         return typ is Any
@@ -27,9 +27,12 @@ try:
     def name_of_generic(typ: Any) -> str:
         return typ._name or typ.__origin__.__name__
 
+    def is_forward_ref(typ: Any) -> bool:
+        return isinstance(typ, ForwardRef)
+
 except ImportError:
     # Python 3.6
-    from typing import _Any, _Union, GenericMeta  # type: ignore
+    from typing import _Any, _Union, GenericMeta, _ForwardRef  # type: ignore
 
     def is_any(typ: Any) -> bool:
         return isinstance(typ, _Any)
@@ -48,3 +51,6 @@ except ImportError:
 
     def name_of_generic(typ: Any) -> str:
         return typ.__name__
+
+    def is_forward_ref(typ: Any) -> bool:
+        return isinstance(typ, _ForwardRef)